careermatch/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.0
python-dotenv==1.0.1
httpx==0.27.0
xmltodict==0.13.0
openai==1.51.0
langchain==0.3.0
langchain-openai==0.2.0
langchain-core==0.3.0
numpy==1.26.4
apscheduler==3.10.4
